refactor(mobile): useRecipientItem
