+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/SwapOrder/hooks/index.ts
@@ -1,1 +1,1 @@
-export { useRecipientItems } from './useRecipientItems'
+export { useRecipientItem } from './useRecipientItem'
